refactor CardDetails component
