+++ apps/backoffice/src/components/cards/index.tsx
@@ -1,2 +1,3 @@
+export * from "./card-rows";
 export * from "./card-details";
 export * from "./card-shortcut";
